requirements.txt
Django>=5.2
python-dotenv>=1.0.0
PyGithub>=2.1.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
celery>=5.3.1
redis>=4.6.0
django-celery-beat>=2.5.0
django-celery-results>=2.6.0
reportlab>=4.4.1
numpy>=2.2.5
scipy>=1.15.3 